Remove SDK fascade
