+++ sdk/src/main/java/com/microsoft/did/sdk/di/SdkComponent.kt
@@ -7,7 +7,9 @@ package com.microsoft.did.sdk.di
 
 import android.content.Context
 import com.microsoft.did.sdk.IdentifierManager
-import com.microsoft.did.sdk.VerifiableCredentialManager
+import com.microsoft.did.sdk.IssuanceService
+import com.microsoft.did.sdk.PresentationService
+import com.microsoft.did.sdk.RevocationService
 import dagger.BindsInstance
 import dagger.Component
 import javax.inject.Named
@@ -27,7 +29,11 @@ internal interface SdkComponent {
 
     fun identifierManager(): IdentifierManager
 
-    fun verifiableCredentialManager(): VerifiableCredentialManager
+    fun issuanceService(): IssuanceService
+
+    fun presentationService(): PresentationService
+
+    fun revocationService(): RevocationService
 
     @Component.Builder
     interface Builder {
