--- sdk/src/test/java/com/microsoft/did/sdk/util/VerifiableCredentialUtilTest.kt
@@ -1,8 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved
-
-package com.microsoft.did.sdk.util
-
-class VerifiableCredentialUtilTest {
-
-
-}
